Square root taylor expansion

+++ IrrationalNumbers/IrrationalNumbers.Logic/IBasicFunctionExpansion.cs
@@ -1,8 +1,10 @@
 ﻿
+using IrrationalNumbers.Core;
+
 namespace IrrationalNumbers.Logic
 {
     public interface IBasicFunctionExpansion
     {
-        double ExpandFunction(int wantedRemainder, double x);
+        BigDecimal ExpandFunction(int wantedRemainder, double x);
     }
 }
